Area de un triangulo y datos a float

+++ prcps-ej/src/Cuadrado.java
@@ -1,9 +1,9 @@
 
 public class Cuadrado extends Figura{
     
-    public Cuadrado(int l1, int l2, int l3, int l4) {
+    public Cuadrado(float l1, float l2, float l3, float l4) {
         super(4);
-        lados = new int[4];
+        lados = new float[4];
         lados[0] = l1;
         lados[1] = l2;
         lados[2] = l3;
