+++ prcps-ej/src/CuadradoCalculadora.java
@@ -8,15 +8,15 @@ public class CuadradoCalculadora implements ICalculadoraFigura{
     }
 
     @Override
-    public int calcularArea() {
-        int[] lados = cuadrado.getLados();
-        int mayor= 0;
+    public float calcularArea() {
+        float[] lados = cuadrado.getLados();
+        float mayor= 0;
         for (int i = 1; i < lados.length; i++) {
             if (lados[i] > mayor) {
 		mayor = lados[i];
             }
 	}
-        int menor= 1000;
+        float menor= 1000;
         for (int i = 1; i < lados.length; i++) {
             if (lados[i] < menor) {
 		menor = lados[i];
@@ -26,9 +26,9 @@ public class CuadradoCalculadora implements ICalculadoraFigura{
     }
 
     @Override
-    public int calcularPerimetro() {
-        int perimetro = 0;
-        int[] lados = cuadrado.getLados();
+    public float calcularPerimetro() {
+        float perimetro = 0;
+        float[] lados = cuadrado.getLados();
         for(int i=0; i<lados.length; i++){
             perimetro+=lados[i];
         }
